+++ packages/device_calendar_plus/test/device_calendar_plus_test.dart
@@ -218,7 +218,8 @@ class MockDeviceCalendarPlusPlatform extends DeviceCalendarPlusPlatform
   }
 
   @override
-  Future<void> showEventModal(String eventId, int? timestamp) async {
+  Future<void> showEventModal(String eventId, int? timestamp,
+      {bool edit = false}) async {
     if (_exceptionToThrow != null) throw _exceptionToThrow!;
   }